[Utils] correcly sort methods between cs_tools and util files
