+++ python/mlp/centroidal/geometric.py
@@ -2,8 +2,7 @@ import numpy as np
 import multicontact_api
 from multicontact_api import ContactPhase, ContactSequence
 from mlp.utils.requirements import Requirements
-from mlp.utils.util import genCOMTrajFromPhaseStates, genAMTrajFromPhaseStates
-from mlp.utils.cs_tools import computePhasesCOMValues
+from mlp.utils.cs_tools import computePhasesCOMValues, genCOMTrajFromPhaseStates, genAMTrajFromPhaseStates
 import math
 import logging
 logging.basicConfig(format='[%(name)-12s] %(levelname)-8s: %(message)s')
